requirements.txt
click
pyperclip
snowflake-connector-python
python-dotenv
openai
sqlfluff
sqlfluff-templater-dbt
wordninja
build
twine
ruamel.yaml<0.18.0
